lang-flashcards/requirements.txt
boto3
fastapi
uvicorn
python-dotenv
openai
gradio
requests